+++ History.md
@@ -7,6 +7,7 @@
 - #931
 - #932
 - #933,  #413, #557, #765
+- make Line Calculation, e.g. total line net amount, accessible via JSON using getCalculation
 
 2.19.0
 =======